📃 docs(quick-start): op download file remind msg

+++ docs/guide/zh-CN/quickstart/startup.sh
@@ -1,9 +1,56 @@
 #!/bin/bash
+# ----------- download relate fiels -----------
+# make sure docker-compose.yml exist
+infraFileLength=`cat docker-compose.yml | wc -l | xargs`
 
-wget -q https://cdn.jsdelivr.net/gh/fBloc/bloc@main/docs/guide/zh-CN/quickstart/docker-compose.yml -O docker-compose.yml
-wget -q https://cdn.jsdelivr.net/gh/fBloc/bloc@main/docs/guide/zh-CN/quickstart/docker-compose-bloc-server-mac.yml -O docker-compose-bloc-server-mac.yml
-wget -q https://cdn.jsdelivr.net/gh/fBloc/bloc@main/docs/guide/zh-CN/quickstart/docker-compose-bloc-server-linux.yml -O docker-compose-bloc-server-linux.yml
-wget -q https://cdn.jsdelivr.net/gh/fBloc/bloc@main/docs/guide/zh-CN/quickstart/shutdown.sh -O shutdown.sh
+if [[ $infraFileLength == *"No such file"* ]]; then  # not exist then download
+	wget -q https://cdn.jsdelivr.net/gh/fBloc/bloc@main/docs/guide/zh-CN/quickstart/docker-compose.yml -O docker-compose.yml
+	infraFileLength=`cat docker-compose.yml | wc -l | xargs`
+fi
+
+if [[ $infraFileLength == "0" ]]; then  # download not valid, reminder user to download handly
+	echo "Download docker-compose.yml fail. plz download https://raw.githubusercontent.com/fBloc/bloc/main/docs/guide/zh-CN/quickstart/docker-compose.yml to your current directory and name it to docker-compose.yml"
+	exit 8
+fi
+
+# make sure docker-compose-bloc-server-mac.yml exist
+macServerComposeFileLineAmount=`cat docker-compose-bloc-server-mac.yml | wc -l | xargs`
+
+if [[ $macServerComposeFileLineAmount == *"No such file"* ]]; then  # not exist then download
+	wget -q https://cdn.jsdelivr.net/gh/fBloc/bloc@main/docs/guide/zh-CN/quickstart/docker-compose-bloc-server-mac.yml -O docker-compose-bloc-server-mac.yml
+	macServerComposeFileLineAmount=`cat docker-compose-bloc-server-mac.yml | wc -l | xargs`
+fi
+
+if [[ $macServerComposeFileLineAmount == "0" ]]; then
+	echo "Download docker-compose-bloc-server-mac.yml fail. plz download https://raw.githubusercontent.com/fBloc/bloc/main/docs/guide/zh-CN/quickstart/docker-compose-bloc-server-mac.yml to your current directory and name it to docker-compose-bloc-server-mac.yml"
+	exit 8
+fi
+
+# make sure docker-compose-bloc-server-linux.yml exist
+linuxServerComposeFileLineAmount=`cat docker-compose-bloc-server-linux.yml | wc -l | xargs`
+if [[ $linuxServerComposeFileLineAmount == *"No such file"* ]]; then  # not exist then download
+	wget -q https://cdn.jsdelivr.net/gh/fBloc/bloc@main/docs/guide/zh-CN/quickstart/docker-compose-bloc-server-linux.yml -O docker-compose-bloc-server-linux.yml
+	linuxServerComposeFileLineAmount=`cat docker-compose-bloc-server-linux.yml | wc -l | xargs`
+fi
+
+if [[ $linuxServerComposeFileLineAmount == "0" ]]; then
+	echo "Download docker-compose-bloc-server-linux.yml fail. plz download https://raw.githubusercontent.com/fBloc/bloc/main/docs/guide/zh-CN/quickstart/docker-compose-bloc-server-linux.yml to your current directory and name it to docker-compose-bloc-server-linux.yml"
+	exit 8
+fi
+
+# make sure shutdown.sh exist
+shutdownShFileLineAmount=`cat shutdown.sh | wc -l | xargs`
+if [[ $shutdownShFileLineAmount == *"No such file"* ]]; then  # not exist then download
+	wget -q https://cdn.jsdelivr.net/gh/fBloc/bloc@main/docs/guide/zh-CN/quickstart/shutdown.sh -O shutdown.sh
+	shutdownShFileLineAmount=`cat shutdown.sh | wc -l | xargs`
+fi
+
+if [[ $shutdownShFileLineAmount == "0" ]]; then
+	echo "Download shutdown.sh fail. plz download https://raw.githubusercontent.com/fBloc/bloc/main/docs/guide/zh-CN/quickstart/shutdown.sh to your current directory and name it to shutdown.sh"
+	exit 8
+fi
+
+# ----------- end download -----------
 
 # check needed port not used
 used_ports=(8083, 8080, 27017, 5672, 15672, 9000, 8086)
